+++ src/main/java/report/model/Employee.java
@@ -1,5 +1,5 @@
 package report.model;
 
-public record Employee(int id, String firstName, String lastName, Long salary, Integer managerId) {
+public record Employee(int id, String firstName, String lastName, Double salary, Integer managerId) {
 }
